// meerkat-core/src/types/duckdb-serialization-types/serialization/Expression.d.ts
export declare enum ExpressionClass {
    INVALID = "INVALID",
    AGGREGATE = "AGGREGATE",
    CASE = "CASE",
    CAST = "CAST",
    COLUMN_REF = "COLUMN_REF",
    COMPARISON = "COMPARISON",
    CONJUNCTION = "CONJUNCTION",
    CONSTANT = "CONSTANT",
    DEFAULT = "DEFAULT",
    FUNCTION = "FUNCTION",
    OPERATOR = "OPERATOR",
    STAR = "STAR",
    SUBQUERY = "SUBQUERY",
    WINDOW = "WINDOW",
    PARAMETER = "PARAMETER",
    COLLATE = "COLLATE",
    LAMBDA = "LAMBDA",
    POSITIONAL_REFERENCE = "POSITIONAL_REFERENCE",
    BETWEEN = "BETWEEN",
    BOUND_AGGREGATE = "BOUND_AGGREGATE",
    BOUND_CASE = "BOUND_CASE",
    BOUND_CAST = "BOUND_CAST",
    BOUND_COLUMN_REF = "BOUND_COLUMN_REF",
    BOUND_COMPARISON = "BOUND_COMPARISON",
    BOUND_CONJUNCTION = "BOUND_CONJUNCTION",
    BOUND_CONSTANT = "BOUND_CONSTANT",
    BOUND_DEFAULT = "BOUND_DEFAULT",
    BOUND_FUNCTION = "BOUND_FUNCTION",
    BOUND_OPERATOR = "BOUND_OPERATOR",
    BOUND_PARAMETER = "BOUND_PARAMETER",
    BOUND_REF = "BOUND_REF",
    BOUND_SUBQUERY = "BOUND_SUBQUERY",
    BOUND_WINDOW = "BOUND_WINDOW",
    BOUND_BETWEEN = "BOUND_BETWEEN",
    BOUND_UNNEST = "BOUND_UNNEST",
    BOUND_LAMBDA = "BOUND_LAMBDA",
    BOUND_LAMBDA_REF = "BOUND_LAMBDA_REF",
    BOUND_EXPRESSION = "BOUND_EXPRESSION"
}